Modification of fuse installation in slice.
Functional "negative" fuses have appeared in the GW5A series - that is,
fuses that are installed if the corresponding attribute/value is not
specified.

Previously, such fuses were only observed in wires, and we took this
into account, but now we have a situation where it is impossible to set
fuses for each slice component separately - "negative" fuses for another
component will be triggered.

So we are moving to a system where all the attributes of all slice
components are collected first, and only then are fuses set with a
complete set of attributes.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>