GW5A. Add ALU and LUTRAM examples.
All examples designed for the GW5A series include the use of ALU, and a
special example of a chain of 4 ALUs is also provided.

To demonstrate the operation of LUTRAM, an example is included using the
ROM16 primitive as the only LUTRAM operating mode supported by the
GW5A-25A and GW5AST-138 chips.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>