Add a mechanism for fully arbitrary ALUs.
If gowin_pack encounters an ALU with the RAW_ALU_LUT parameter set, the
value of this parameter is used to set the ALU fuse without changes.

This allows, for example, in nextpnr, to disable constant inputs from
networks and modify the ALU LUT so that these inputs do not affect the
result.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>