GW5A. Masquerade for memory cells.
Since only ROM16 is supported from the "normal" LUTRAM modes in the GW5A
series, we do not let nextpnr think that we have full memory. We mask
the cells as normal logic.

This temporary measure will work well until we have to implement the 60K
chip option, at which point we will change nextpnr.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>