GW5A. Add OSC (#379)
* GW5A. Add OSC

A programmable on-chip crystal oscillator (OSCA) has been implemented for GW5A-25A.

A distinctive feature of the new series is that the OSC output is no
longer a conventional wire, and therefore routing it through
conventional PIPs is not possible, nor is it possible to "pull" its
signal to the entry point of the global clock wire system.

Nextpnr has been modified to connect the OSC output via an intermediate
clock wire.

In addition, the OSC fuses are also "spread" across the chip like clock
fuses.

* GW5A. OSC divider 3

In addition to even divisors, OSC in the GW5A series also supports
division by 3.

Let's take this into account.

---------

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>