Fix ALU
When CARRY is generated in logic, nextpnr v0.9 outputs incorrect
parameters for the first ALU in the chain.

A fix for nextpnr is already in the works, but it will only be included
in the next release, so we are adding a workaround to apicula to support
both 0.9 and subsequent releases of nextpnr.

---------

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>