Document the slots.
In the GW5A series, some fairly large functional blocks are located
outside the main grid—that is, the grid has wires for pins, almost
always scattered across several cells, but there are no functional
fuses—these fuses are installed in special blocks (slots) that are not
geometrically related to the main grid.

Fuses are described in tables with types >1024, although there are
naturally no tiles of this type, since there is no place for them in the
main grid.

Each slot has a unique number that determines which primitive is located
in it.

In fact, it is quite easy to determine the purpose of the slots—the
output file contains blocks describing the slots used. So, if the lower
left PLL is used (Gowin IDE allows you to specify the desired location
of primitives), then the output file will contain a section describing
slot number 6.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>