PLL in a separate module. (#394)
The GW5A series has a different PLL, so we will adapt our examples to
future support for the GW5A series.

TangPrimer25k does not have an LCD VGA connector, so we will use a VGA
PMOD, which will require the use of a normal monitor that will probably
not support 420x272.

Different boards have different numbers of wires connected to the VGA
connector (not to mention that the TangPrimer25k does not have this
connector at all and requires the use of PMOD).
We take this fact into account in the examples.

---------

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>