GW5. Implement ADC.
ADC implementation for GW5A-25 chips is added.

Parameters such as clock frequency divider, clock source, sample rate,
sample count, scale, offset, and others are set.

Pins used as a signal source for the ADC are switched to a special
operating mode. At the same time, the limitations imposed by the Gowin
ADC are monitored - only one pair of IOs per bus from a specific list of
IO banks.

ADC bus 2 is supported, which provides dynamic switching between bank 1
pins.

IO TLVDS_IBUF_ADC primitives are implemented, which provide a signal for
bus 2.

The 12-bit offset can be either negative or positive, so we take this
fact into account.

An example of using ADC to measure the internal temperature of
TangPrimer25k and display it on LED PMODs is added.

Dynamically switch between two signal sources for the ADC.

Use bus1 pin as an input for ADC.

Note that this is likely to be a very noisy channel – both Gowin IDE and
Apicula binary streams can produce false readings of up to 0V.


Signed-off-by: YRabbit <rabbit@yrabbit.cyou>