Read information about banks from a DAT file.
We are changing the source of information about the banks to which IOs
belong. Previously, this information was obtained from a specific
package, but now it will be taken from the DAT file that describes the
chip.

This will allow you to work with pins that are not exposed in any
package. You cannot do much with them, but it is advisable to set them
to ‘not used’ status.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>