Read a list of alt IO functions from a DAT file.
The IO buffer can be linked to several internal mechanisms and perform
functions other than GPIO.

The method currently used to detect these alternative IO pin functions
is not entirely reliable. It is based on the configuration of a specific
package, but the fact is that not all pins are used even in the largest
package.

The DAT file section, on the other hand, contains all the information,
but not in text form.

We use a combined approach - where possible, we take the description
from the package, in other cases from the DAT.

Signed-off-by: YRabbit <rabbit@yrabbit.cyou>